emove "actions/" and moved logic into utils/api
