--- src/actions/setup/createProfile.ts
@@ -1,11 +1,0 @@
-"use server";
-
-import { db } from "@/db";
-import { InsertProfileRecord, profiles } from "@/db/schema";
-
-export async function createProfile({ userName, id }: InsertProfileRecord) {
-  return await db.insert(profiles).values({
-    id: id,
-    userName: userName,
-  });
-}
