--- src/actions/setup/index.ts
@@ -1,2 +1,0 @@
-export * from "./createProfile";
-export * from "./createTargetKcal";
