+++ src/app/auth/_components/ButtonWithGooleIcon.tsx
@@ -10,6 +10,7 @@ const Component = ({ text, submitGoogle }: ButtonWithGooleIconProps) => {
   return (
     <div className="relative z-10">
       <button
+        type="button"
         onClick={submitGoogle}
         className={cn(
           "w-full bg-card text-card-foreground flex items-center justify-center rounded-lg border-1 border-foreground gap-2 px-4 py-2 md:text-sm"
